refactor(add): split cli add shared helpers
