+++ packages/wp-typia-project-tools/tests/cli-add-shared.test.ts
@@ -7,10 +7,54 @@ import {
   assertScaffoldDoesNotExist,
 } from '../src/runtime/cli-add-shared.js';
 
+const runtimeRoot = path.join(import.meta.dir, '..', 'src', 'runtime');
 const tempRoot = fs.mkdtempSync(path.join(os.tmpdir(), 'wp-typia-add-shared-'));
 
 afterAll(() => {
   fs.rmSync(tempRoot, { force: true, recursive: true });
+});
+
+function readRuntimeSource(fileName: string): string {
+  return fs.readFileSync(path.join(runtimeRoot, fileName), 'utf8');
+}
+
+test('shared add runtime keeps compatibility exports around focused modules', () => {
+  const sharedSource = readRuntimeSource('cli-add-shared.ts');
+
+  expect(sharedSource).toContain('export * from "./cli-add-types.js";');
+  expect(sharedSource).toContain('export * from "./cli-add-validation.js";');
+  expect(sharedSource).toContain('export * from "./cli-add-filesystem.js";');
+  expect(sharedSource).toContain('export * from "./cli-add-block-json.js";');
+  expect(sharedSource).toContain('export * from "./cli-add-collision.js";');
+  expect(sharedSource).toContain('export * from "./cli-add-help.js";');
+  expect(sharedSource).toContain('from "./scaffold-identifiers.js"');
+  expect(sharedSource).not.toContain('export interface RunAddBlockCommandOptions');
+  expect(sharedSource).not.toContain('export function assertValidGeneratedSlug');
+  expect(sharedSource).not.toContain('export async function patchFile');
+  expect(sharedSource).not.toContain('export function assertScaffoldDoesNotExist');
+  expect(sharedSource).not.toContain('export function formatAddHelpText');
+});
+
+test('focused add runtime modules own their helper categories', () => {
+  const typesSource = readRuntimeSource('cli-add-types.ts');
+  const validationSource = readRuntimeSource('cli-add-validation.ts');
+  const filesystemSource = readRuntimeSource('cli-add-filesystem.ts');
+  const blockJsonSource = readRuntimeSource('cli-add-block-json.ts');
+  const collisionSource = readRuntimeSource('cli-add-collision.ts');
+  const helpSource = readRuntimeSource('cli-add-help.ts');
+
+  expect(typesSource).toContain('export interface RunAddBlockCommandOptions');
+  expect(typesSource).toContain('export const ADD_KIND_IDS');
+  expect(validationSource).toContain('export function assertValidGeneratedSlug');
+  expect(validationSource).toContain('export function assertValidRestResourceMethods');
+  expect(validationSource).toContain('export function assertValidEditorPluginSlot');
+  expect(filesystemSource).toContain('export async function snapshotWorkspaceFiles');
+  expect(filesystemSource).toContain('export async function rollbackWorkspaceMutation');
+  expect(blockJsonSource).toContain('export function readWorkspaceBlockJson');
+  expect(blockJsonSource).toContain('export function getMutableBlockHooks');
+  expect(collisionSource).toContain('export function assertScaffoldDoesNotExist');
+  expect(collisionSource).toContain('export function assertEditorPluginDoesNotExist');
+  expect(helpSource).toContain('export function formatAddHelpText');
 });
 
 test('shared add collision helper allows missing filesystem paths and inventory entries', () => {
