+++ packages/wp-typia/src/print-line.ts
@@ -1,0 +1,1 @@
+export type PrintLine = (line: string) => void;
